all data now sends to the new producers, except most_used_words

+++ src/main/kotlin/offtop/sessionReportService/Models/EndSessionReport.kt
@@ -3,7 +3,7 @@ package offtop.sessionReportService.Models
 class EndSessionReport constructor(
         userId: Int,
         sessionData: List<SessionData>,
-        averageFocusScore: Int,
+        averageFocusScore: Double,
         topic: String,
         mostUsedWord: List<String>,
         startTime: String,
@@ -12,7 +12,7 @@ class EndSessionReport constructor(
 ) {
     var user_id: Int = userId
     var session_data: List<SessionData> = sessionData
-    var average_focus_score: Int = averageFocusScore
+    var average_focus_score: Double = averageFocusScore
     var topic: String = topic
     var most_used_word: List<String> = mostUsedWord
     var start_time: String = startTime
